ocr/requirements.txt
torch
transformers
pillow
flask
requests
pdf2image
pytesseract